refactor: 깃허브 피드백 반영
