+++ app/(main)/customer/cart/[id]/_types/CartCardType.ts
